Split .hpp into declaratation (.hpp) and definition (.cpp)
