+++ src/lib/AAMutSelNeCodonMatrixBidimArray.hpp
@@ -1,10 +1,7 @@
 #pragma once
 
-#include <cassert>
 #include "AAMutSelOmegaCodonSubMatrix.hpp"
-#include "Array.hpp"
 #include "BidimArray.hpp"
-#include "IIDDirichlet.hpp"
 
 /**
  * \brief A BidimArray of AAMutSelOmegaCodonSubMatrix
@@ -23,27 +20,9 @@ class MutSelNeCodonMatrixBidimArray : public BidimArray<SubMatrix>,
     //! fitness profiles and an array of population size.
     MutSelNeCodonMatrixBidimArray(const CodonStateSpace *incodonstatespace,
         const SubMatrix *innucmatrix, const Selector<std::vector<double>> *infitnessarray,
-        const std::vector<double> &pop_size_array)
-        : codonstatespace(incodonstatespace),
-          nucmatrix(innucmatrix),
-          aafitnessarray(infitnessarray),
-          matrixbidimarray(pop_size_array.size(),
-              std::vector<AAMutSelOmegaCodonSubMatrix *>(infitnessarray->GetSize(), nullptr)) {
-        std::cout << GetNrow() << "\t" << GetNcol() << "\n";
-        for (int i = 0; i < GetNrow(); i++) {
-            for (int j = 0; j < GetNcol(); j++) {
-                matrixbidimarray[i][j] = new AAMutSelOmegaCodonSubMatrix(codonstatespace, nucmatrix,
-                    infitnessarray->GetVal(j), 1.0, pop_size_array.at(i));
-            }
-        }
-        assert(static_cast<int>(matrixbidimarray.size()) == GetNrow());
-    }
+        const std::vector<double> &pop_size_array);
 
-    ~MutSelNeCodonMatrixBidimArray() override {
-        for (int i = 0; i < GetNrow(); i++) {
-            for (int j = 0; j < GetNcol(); j++) { delete matrixbidimarray[i][j]; }
-        }
-    }
+    ~MutSelNeCodonMatrixBidimArray() override;
 
     //! return the number of rows (number of branches)
     int GetNrow() const override { return matrixbidimarray.size(); }
@@ -51,61 +30,30 @@ class MutSelNeCodonMatrixBidimArray : public BidimArray<SubMatrix>,
     int GetNcol() const override { return aafitnessarray->GetSize(); }
 
     //! const access to the matrix for row (branch) i and column (site) j
-    const AAMutSelOmegaCodonSubMatrix &GetVal(int i, int j) const override {
-        assert(0 <= i and i < GetNrow());
-        assert(0 <= j and j < GetNcol());
-        return *matrixbidimarray.at(i).at(j);
-    }
+    const AAMutSelOmegaCodonSubMatrix &GetVal(int i, int j) const override;
 
     //! non const access to the matrix for row (branch) i and column (site) j
-    AAMutSelOmegaCodonSubMatrix &operator()(int i, int j) override {
-        assert(0 <= i and i < GetNrow());
-        assert(0 <= j and j < GetNcol());
-        return *matrixbidimarray.at(i).at(j);
-    }
+    AAMutSelOmegaCodonSubMatrix &operator()(int i, int j) override;
 
     //! Set ne for row (branch) i
-    void SetRowNe(int i, double Ne) {
-        for (int j = 0; j < this->GetNcol(); j++) { (*this)(i, j).SetNe(Ne); }
-    }
+    void SetRowNe(int i, double Ne);
 
-    void SetNe(std::vector<double> const &Ne) {
-        assert(GetNcol() > 0);
-        assert(GetNrow() > 0);
-        assert(GetNrow() == static_cast<int>(Ne.size()));
-        for (int i = 0; i < this->GetNrow(); i++) {
-            SetRowNe(i, Ne[i]);
-        }
-    }
+    void SetNe(std::vector<double> const &Ne);
 
     //! signal corruption of the parameters (matrices should recompute themselves)
-    void CorruptCodonMatrices() {
-        for (int j = 0; j < GetNcol(); j++) { CorruptColCodonMatrices(j); }
-    }
+    void CorruptCodonMatrices();
 
     //! signal corruption for column (site) j
-    void CorruptColCodonMatrices(int j) {
-        assert(GetNrow() > 0);
-        for (int i = 0; i < GetNrow(); i++) { CorruptMatrix(i, j); }
-    }
+    void CorruptColCodonMatrices(int j);
 
     //! signal corruption for row (branch) i
-    void CorruptRowCodonMatrices(int i) {
-        assert(GetNcol() > 0);
-        for (int j = 0; j < GetNcol(); j++) { CorruptMatrix(i, j); }
-    }
+    void CorruptRowCodonMatrices(int i);
 
     //! signal corruption for row (branch) i and column (site) j
-    void CorruptMatrix(int i, int j) {
-        (*this)(i, j).CorruptMatrix();
-    }
+    void CorruptMatrix(int i, int j);
 
     //! update only those matrices for which occupancy[i] != 0
-    void UpdateCodonMatrices(const Selector<int> &occupancy) {
-        for (int col = 0; col < GetNcol(); col++) {
-            if (!occupancy.GetVal(col)) { (*this).CorruptColCodonMatrices(col); }
-        }
-    }
+    void UpdateCodonMatrices(const Selector<int> &occupancy);
 
   private:
     const CodonStateSpace *codonstatespace;
@@ -129,14 +77,7 @@ class AAMutSelNeCodonSubMatrixArray : public Array<SubMatrix>,
     //! profiles and a single Ne value (for all matrices)
     AAMutSelNeCodonSubMatrixArray(const CodonStateSpace *incodonstatespace,
         const SubMatrix *innucmatrix, const Selector<std::vector<double>> *inaafitnessarray,
-        double inne)
-        : codonstatespace(incodonstatespace),
-          nucmatrix(innucmatrix),
-          aafitnessarray(inaafitnessarray),
-          ne(inne),
-          matrixarray(inaafitnessarray->GetSize()) {
-        Create();
-    }
+        double inne);
 
     ~AAMutSelNeCodonSubMatrixArray() { Delete(); }
 
@@ -157,27 +98,15 @@ class AAMutSelNeCodonSubMatrixArray : public Array<SubMatrix>,
     const SubMatrix &GetNucMatrix() const { return *nucmatrix; }
 
     //! update all matrices
-    void UpdateCodonMatrices() {
-        for (int i = 0; i < GetSize(); i++) { UpdateCodonMatrices(i); }
-    }
+    void UpdateCodonMatrices();
 
     //! update matrices of component i
-    void UpdateCodonMatrices(int i) {
-        (*this)[i].SetNe(ne);
-        (*this)[i].CorruptMatrix();
-    }
+    void UpdateCodonMatrices(int i);
 
   private:
-    void Create() {
-        for (int i = 0; i < GetSize(); i++) {
-            matrixarray[i] = new AAMutSelOmegaCodonSubMatrix(
-                codonstatespace, nucmatrix, aafitnessarray->GetVal(i), 1.0, ne);
-        }
-    }
+    void Create();
 
-    void Delete() {
-        for (int i = 0; i < GetSize(); i++) { delete matrixarray[i]; }
-    }
+    void Delete();
 
     const CodonStateSpace *codonstatespace;
     const SubMatrix *nucmatrix;
