+++ src/lib/SuffStat.hpp
@@ -1,7 +1,4 @@
-
-#ifndef SUFFSTAT_H
-#define SUFFSTAT_H
-
+#pragma once
 /**
  * \brief A dummy superclass for all sufficient statistics. Does not implement
  * anything in itself and does not provide any specific interface.
@@ -12,5 +9,3 @@ class SuffStat {
     SuffStat() {}
     virtual ~SuffStat() {}
 };
-
-#endif
